sessions

+++ src/Manager/SessionManager.php
@@ -11,12 +11,32 @@
 namespace App\Manager;
 
 use Doctrine\DBAL\DBALException;
+use Doctrine\ORM\EntityManagerInterface;
+use SessionHandlerInterface;
+use Symfony\Component\HttpFoundation\Session\Storage\Handler\PdoSessionHandler;
 
 /**
  * Class SessionManager.
  */
 class SessionManager extends AbstractManager
 {
+    /**
+     * @var SessionHandlerInterface
+     */
+    private $sessionHandler;
+
+    /**
+     * SessionManager constructor.
+     *
+     * @param EntityManagerInterface  $entityManager
+     * @param SessionHandlerInterface $sessionHandler
+     */
+    public function __construct(EntityManagerInterface $entityManager, SessionHandlerInterface $sessionHandler)
+    {
+        parent::__construct($entityManager);
+        $this->sessionHandler = $sessionHandler;
+    }
+
     /**
      * @param int $maxlifetime
      *
@@ -28,12 +48,17 @@ class SessionManager extends AbstractManager
      */
     public function purge(int $maxlifetime): self
     {
-        $query = 'DELETE FROM `sessions` WHERE `sess_time` < :time';
+        /*
+         * ONLY AVAILABLE FOR PDO SESSION HANDLER
+         */
+        if ($this->sessionHandler instanceof PdoSessionHandler) {
+            $query = 'DELETE FROM `sessions` WHERE `sess_time` < :time';
 
-        $statement = $this->entityManager->getConnection()->prepare($query);
+            $statement = $this->entityManager->getConnection()->prepare($query);
 
-        $statement->bindValue(':time', time() - $maxlifetime);
-        $statement->execute();
+            $statement->bindValue(':time', time() - $maxlifetime);
+            $statement->execute();
+        }
 
         return $this;
     }
